fixed numerous styling issues and refactored code
