--- common/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-anyhow = "1.0.98"
-crc = "3.2.1"
