--- common/src/modbus/mod.rs
@@ -1,1 +1,0 @@
-pub mod modbus;
