+++ control-core/Cargo.toml
@@ -15,7 +15,6 @@ serde = "1.0.219"
 serde_json = "1.0.140"
 smol = "2.0.2"
 socketioxide = "0.16.2"
-common = { path = "../common"}
 uom = "0.36.0"
 bitvec = { version = "1.0.1", features = ["alloc"] }
 crc = "3.2.1"
