+++ control-core/src/lib.rs
@@ -4,5 +4,6 @@ pub mod converters;
 pub mod ethercat;
 pub mod identification;
 pub mod machines;
+pub mod modbus;
 pub mod rest;
 pub mod socketio;
