+++ control-core/src/modbus/mod.rs
@@ -7,116 +7,11 @@ use std::{
 use anyhow::Error;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SerialEncoding {
-    Coding7E1, // 7 data, even parity, 1 stop
-    Coding7O1, // 7 data, odd parity, 1 stop
-    Coding7E2, // 7 data, even parity, 2 stop
-    Coding7O2, // 7 data, odd parity, 2 stop
-    Coding8N1, // 8 data, no parity, 1 stop
-    Coding8E1, // 8 data, even parity, 1 stop
-    Coding8O1, // 8 data, odd parity, 1 stop
-    Coding8N2, // 8 data, no parity, 2 stop
-    Coding8E2, // 8 data, even parity, 2 stop
-    Coding8O2, // 8 data, odd parity, 2 stop
-    Coding8S1, // 8 data, space parity, 1 stop
-    Coding8M1, // 8 data, mark parity, 1 stop
-}
-
-impl SerialEncoding {
-    /// Get the number of data bits
-    pub fn data_bits(&self) -> u8 {
-        match self {
-            Self::Coding7E1 | Self::Coding7O1 | Self::Coding7E2 | Self::Coding7O2 => 7,
-            _ => 8,
-        }
-    }
-
-    /// Get the number of parity bits (0 or 1)
-    pub fn parity_bits(&self) -> u8 {
-        match self {
-            Self::Coding8N1 | Self::Coding8N2 => 0,
-            _ => 1,
-        }
-    }
-
-    /// Get the parity type
-    pub fn parity_type(&self) -> Option<ParityType> {
-        match self {
-            Self::Coding7E1 | Self::Coding7E2 | Self::Coding8E1 | Self::Coding8E2 => {
-                Some(ParityType::Even)
-            }
-            Self::Coding7O1 | Self::Coding7O2 | Self::Coding8O1 | Self::Coding8O2 => {
-                Some(ParityType::Odd)
-            }
-            Self::Coding8S1 => Some(ParityType::Space),
-            Self::Coding8M1 => Some(ParityType::Mark),
-            Self::Coding8N1 | Self::Coding8N2 => None,
-        }
-    }
-
-    /// Get the number of stop bits
-    pub fn stop_bits(&self) -> u8 {
-        match self {
-            Self::Coding7E1
-            | Self::Coding7O1
-            | Self::Coding8N1
-            | Self::Coding8E1
-            | Self::Coding8O1
-            | Self::Coding8S1
-            | Self::Coding8M1 => 1,
-            Self::Coding7E2
-            | Self::Coding7O2
-            | Self::Coding8N2
-            | Self::Coding8E2
-            | Self::Coding8O2 => 2,
-        }
-    }
-
-    /// Get the total number of bits sent per byte according to the SerialEncoding (including start bit)
-    /// For Example: With 8n1 transferring 1 byte over Serial actually transfers 10 bits -> 8 data bits, 0 parity, 1 start bit and 1 stop bit
-    pub fn total_bits(&self) -> u8 {
-        // Start bit is always 1
-        1 + self.data_bits() + self.parity_bits() + self.stop_bits()
-    }
-
-    /// Get a human-readable description
-    pub fn description(&self) -> String {
-        let data = self.data_bits();
-        let parity = match self.parity_type() {
-            Some(ParityType::Even) => "E",
-            Some(ParityType::Odd) => "O",
-            Some(ParityType::Space) => "S",
-            Some(ParityType::Mark) => "M",
-            None => "N",
-        };
-        let stop = self.stop_bits();
-
-        format!(
-            "{}{}{}({} bits total)",
-            data,
-            parity,
-            stop,
-            self.total_bits()
-        )
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ParityType {
     Even,
     Odd,
     Space,
     Mark,
-}
-
-fn convert_nanoseconds_to_milliseconds(nanoseconds: u64) -> u128 {
-    let duration: Duration = Duration::from_nanos(nanoseconds);
-    return duration.as_millis();
-}
-
-fn convert_milliseconds_to_nanoseconds(milliseconds: u64) -> u128 {
-    let duration: Duration = Duration::from_millis(milliseconds);
-    return duration.as_nanos();
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -144,6 +39,24 @@ pub struct ModbusResponse {
     pub function_code: ModbusFunctionCode, // needs to be u8 because of exceptions
     pub data: Vec<u8>,
     pub crc: u16,
+}
+
+impl ModbusResponse {
+    fn is_exception(self) -> bool {
+        if self.data.len() > 2 {
+            let function_code = self.data[1];
+            return (function_code & 0b10000000) > 0; // first bit is set when an exception happens in Modbus
+        }
+        return false;
+    }
+
+    fn get_exception_code(self) -> u8 {
+        if self.data.len() > 3 {
+            let exception = self.data[2];
+            return exception;
+        }
+        return 0;
+    }
 }
 
 impl From<u8> for ModbusFunctionCode {
